Refact: 메서드 위치 변경
