+++ src/main/java/com/est/jobshin/domain/interview/controller/InterviewController.java
@@ -2,9 +2,9 @@ package com.est.jobshin.domain.interview.controller;
 
 import com.est.jobshin.domain.interview.domain.Interview;
 import com.est.jobshin.domain.interview.dto.InterviewDto;
-import com.est.jobshin.domain.interview.service.InterviewResultService;
 import com.est.jobshin.domain.interview.service.InterviewService;
 import com.est.jobshin.domain.interviewDetail.domain.InterviewDetail;
+import com.est.jobshin.domain.interviewDetail.service.InterviewDetailService;
 import com.est.jobshin.infra.alan.AlanService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -17,8 +17,8 @@ import org.springframework.web.bind.annotation.*;
 public class InterviewController {
 
     private final InterviewService interviewService;
-    private final InterviewResultService interviewResultService;
     private final AlanService alanService;
+    private final InterviewDetailService interviewDetailService;
 
     @GetMapping("/test")
     public String callAlan() {
@@ -49,14 +49,9 @@ public class InterviewController {
 
     @GetMapping("/{interviewDetailId}/userAnswer")
     public ResponseEntity<InterviewDetail> saveAnswer(@PathVariable("interviewDetailId") Long interviewDetailId, String userAnswer) {
-        interviewResultService.saveAnswer(interviewDetailId, "userAnswer");
-        InterviewDetail interviewDetail = interviewResultService.createExampleAnswer(interviewDetailId, "userAnswer");
+        interviewDetailService.saveAnswer(interviewDetailId, "userAnswer");
+        InterviewDetail interviewDetail = interviewDetailService.createExampleAnswer(interviewDetailId, "userAnswer");
         return ResponseEntity.ok(interviewDetail);
-    }
-
-    @GetMapping("/{interviewId}/{interviewDetailId}")
-    public ResponseEntity<InterviewDetail> getInterviewDetail(@PathVariable("interviewId") Long interviewId, @PathVariable("interviewDetailId") Long interviewDetailId) {
-        return ResponseEntity.ok(interviewResultService.getInterviewDetail(interviewDetailId));
     }
 
     @GetMapping("/{interviewId}")
